unlatch if jump
If the player tries to jump while latched, they get unlatched and the anchor gets pulled with them.
There's a strange interaction of if grounded and jumping long enough, the anchor will detach, will require a bit of work to change that to ignore any jump input if was grounded when jumped, but don't know if necessary.